Read the desc, cause its long
Okay so, the player needed to know how much it costed to traverse from one node to the other.
I found a way to label them, the label appears in the midpoint between the two nodes.
However, the numbers that are generated are weird and should probably be fixed, or understood