MOB UPDATE & CASTLE TILES
As title states this is the long awaited, much hyped, mob update. We now have the first enemy npc in the game and it could not be more beautiful. The enemy "Castle Rat" does not have any attack animations...we are safe...for now.

There is also a tileset designed to look like a castle.